Converted view to use SolidGrid and build-in tile creation, moved grid view update part to Maze.kt
